+++ packages/plugin-e2e/src/fixtures/commands/createDataSource.ts
@@ -1,12 +1,10 @@
 import { v4 as uuidv4 } from 'uuid';
 import { APIRequestContext, expect, TestFixture } from '@playwright/test';
-import { PluginFixture, PluginOptions } from '../../api';
-import { CreateDataSourceArgs, DataSourceSettings } from '../../types';
-import { PlaywrightCombinedArgs } from '../types';
+import { CreateDataSourceArgs, DataSourceSettings, PlaywrightArgs } from '../../types';
 
 type CreateDataSourceViaAPIFixture = TestFixture<
   (args: CreateDataSourceArgs) => Promise<DataSourceSettings>,
-  PluginFixture & PluginOptions & PlaywrightCombinedArgs
+  PlaywrightArgs
 >;
 
 export const createDataSourceViaAPI = async (
@@ -40,10 +38,8 @@ export const createDataSourceViaAPI = async (
   return existingDataSource.json();
 };
 
-const createDataSource: CreateDataSourceViaAPIFixture = async ({ request }, use) => {
+export const createDataSource: CreateDataSourceViaAPIFixture = async ({ request }, use) => {
   await use(async (args) => {
     return createDataSourceViaAPI(request, args);
   });
 };
-
-export default createDataSource;