+++ packages/plugin-e2e/src/fixtures/commands/login.ts
@@ -1,15 +1,12 @@
 import path from 'path';
 import { expect, TestFixture } from '@playwright/test';
-import { PluginFixture, PluginOptions } from '../../api';
-import { PlaywrightCombinedArgs } from '../types';
+import { PlaywrightArgs } from '../../types';
 
-const authFile = 'playwright/.auth/user.json';
-
-type LoginFixture = TestFixture<() => Promise<void>, PluginFixture & PluginOptions & PlaywrightCombinedArgs>;
+type LoginFixture = TestFixture<() => Promise<void>, PlaywrightArgs>;
 
 const ADMIN_USER = { user: 'admin', password: 'admin' };
 
-const login: LoginFixture = async ({ request, user }, use) => {
+export const login: LoginFixture = async ({ request, user }, use) => {
   await use(async () => {
     const data = user ?? ADMIN_USER;
     const loginReq = await request.post('/login', { data });
@@ -18,5 +15,3 @@ const login: LoginFixture = async ({ request, user }, use) => {
     await request.storageState({ path: path.join(process.cwd(), `playwright/.auth/${data.user}.json`) });
   });
 };
-
-export default login;