+++ packages/plugin-e2e/src/fixtures/isFeatureToggleEnabled.ts
@@ -1,17 +1,11 @@
 import { TestFixture } from '@playwright/test';
-import { PluginFixture, PluginOptions } from '../api';
-import { PlaywrightCombinedArgs } from './types';
+import { PlaywrightArgs } from '../types';
 
-type FeatureToggleFixture = TestFixture<
-  <T = object>(featureToggle: keyof T) => Promise<boolean>,
-  PluginFixture & PluginOptions & PlaywrightCombinedArgs
->;
+type FeatureToggleFixture = TestFixture<<T = object>(featureToggle: keyof T) => Promise<boolean>, PlaywrightArgs>;
 
-const isFeatureToggleEnabled: FeatureToggleFixture = async ({ page, grafanaVersion }, use) => {
+export const isFeatureToggleEnabled: FeatureToggleFixture = async ({ page, grafanaVersion }, use) => {
   await use(async <T = object>(featureToggle: keyof T) => {
     const featureToggles: T = await page.evaluate('window.grafanaBootData.settings.featureToggles');
     return Boolean(featureToggles[featureToggle]);
   });
 };
-
-export default isFeatureToggleEnabled;